requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
async-generator==1.10
azure-core==1.13.0
azure-storage-blob==12.8.0
backcall==0.2.0
cdifflib==1.2.5
certifi==2020.12.5
chardet==4.0.0
confuse==1.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.10
django-cors-headers==3.7.0
django-extensions==3.1.1
django-mssql-backend==2.8.1
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.12.3
djangorestframework-simplejwt
entrypoints==0.3
et-xmlfile==1.0.1
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
ipywidgets==7.6.3
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.2
joblib==1.0.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.35.0
matplotlib==3.3.3
missingno==0.4.2
msrest==0.6.21
mysql-connector-python==8.0.24
networkx==2.5
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.7
pandas
pandas-profiling==2.10.0
parso==0.8.1
pdfminer.six==20200517
pdfplumber==0.5.28
phik==0.11.0
Pillow==8.1.0
protobuf==3.16.0
pycryptodome==3.10.1
PyJWT==2.0.1
pymssql==2.1.5
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
scipy
seaborn
simplejson==3.17.2
sortedcontainers==2.4.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
tangled-up-in-unicode==0.0.6
terminado==0.9.2
testpath==0.4.4
tqdm==4.56.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
visions==0.6.0
Wand==0.6.6
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1

